creating table search POM based on review comments
